# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Class Routine Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TimetableEntryDayOfWeek =
  (typeof TimetableEntryDayOfWeek)[keyof typeof TimetableEntryDayOfWeek];

export const TimetableEntryDayOfWeek = {
  Monday: "Monday",
  Tuesday: "Tuesday",
  Wednesday: "Wednesday",
  Thursday: "Thursday",
  Friday: "Friday",
  Saturday: "Saturday",
  Sunday: "Sunday",
} as const;

export interface TimetableEntry {
  id: number;
  courseName: string;
  courseCode?: string;
  instructor?: string;
  dayOfWeek: TimetableEntryDayOfWeek;
  startTime: string;
  endTime: string;
  room?: string;
  color?: string;
}

export type CreateTimetableEntryDayOfWeek =
  (typeof CreateTimetableEntryDayOfWeek)[keyof typeof CreateTimetableEntryDayOfWeek];

export const CreateTimetableEntryDayOfWeek = {
  Monday: "Monday",
  Tuesday: "Tuesday",
  Wednesday: "Wednesday",
  Thursday: "Thursday",
  Friday: "Friday",
  Saturday: "Saturday",
  Sunday: "Sunday",
} as const;

export interface CreateTimetableEntry {
  courseName: string;
  courseCode?: string;
  instructor?: string;
  dayOfWeek: CreateTimetableEntryDayOfWeek;
  startTime: string;
  endTime: string;
  room?: string;
  color?: string;
}

export type AssignmentStatus =
  (typeof AssignmentStatus)[keyof typeof AssignmentStatus];

export const AssignmentStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type AssignmentPriority =
  (typeof AssignmentPriority)[keyof typeof AssignmentPriority];

export const AssignmentPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Assignment {
  id: number;
  title: string;
  courseName: string;
  dueDate: string;
  description?: string;
  status: AssignmentStatus;
  priority: AssignmentPriority;
}

export type CreateAssignmentStatus =
  (typeof CreateAssignmentStatus)[keyof typeof CreateAssignmentStatus];

export const CreateAssignmentStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type CreateAssignmentPriority =
  (typeof CreateAssignmentPriority)[keyof typeof CreateAssignmentPriority];

export const CreateAssignmentPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateAssignment {
  title: string;
  courseName: string;
  dueDate: string;
  description?: string;
  status: CreateAssignmentStatus;
  priority: CreateAssignmentPriority;
}

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType];

export const ResourceType = {
  github: "github",
  colab: "colab",
  docs: "docs",
  other: "other",
} as const;

export interface Resource {
  id: number;
  title: string;
  url: string;
  type: ResourceType;
  courseName?: string;
  description?: string;
}

export type CreateResourceType =
  (typeof CreateResourceType)[keyof typeof CreateResourceType];

export const CreateResourceType = {
  github: "github",
  colab: "colab",
  docs: "docs",
  other: "other",
} as const;

export interface CreateResource {
  title: string;
  url: string;
  type: CreateResourceType;
  courseName?: string;
  description?: string;
}