# lib/api-zod/src/generated/types/createAssignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Class Routine Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAssignmentPriority } from "./createAssignmentPriority";
import type { CreateAssignmentStatus } from "./createAssignmentStatus";

export interface CreateAssignment {
  title: string;
  courseName: string;
  dueDate: Date;
  description?: string;
  status: CreateAssignmentStatus;
  priority: CreateAssignmentPriority;
}