# lib/api-zod/src/generated/types/createResource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Class Routine Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResourceType } from "./createResourceType";

export interface CreateResource {
  title: string;
  url: string;
  type: CreateResourceType;
  courseName?: string;
  description?: string;
}