# lib/api-zod/src/generated/types/createTimetableEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Class Routine Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTimetableEntryDayOfWeek } from "./createTimetableEntryDayOfWeek";

export interface CreateTimetableEntry {
  courseName: string;
  courseCode?: string;
  instructor?: string;
  dayOfWeek: CreateTimetableEntryDayOfWeek;
  startTime: string;
  endTime: string;
  room?: string;
  color?: string;
}