# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Class Routine Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from "./resourceType";

export interface Resource {
  id: number;
  title: string;
  url: string;
  type: ResourceType;
  courseName?: string;
  description?: string;
}